Make sure configuration is saved to storage device
Config file is written to a temp file and then it is renamed to the
original config file. However, it is possible that the rename operation
will be commited to storage while file data will be still in cache
causing original config file to be empty or partially written in case of
a system reboot without a clean shutdown. Make this less likely to occur
by forcing the data to be written to the storage device before renaming
the file.

Signed-off-by: Dmitry Shmidt <dimitrysh@google.com>

Git-commit: a218e1ded44cc88072f4e1896fdd8bd24f1829ac
Git-repo: git://w1.fi/srv/git/hostap.git
Change-Id: I8243f724af20eef8d2684dc91a0b9569e4509d0e
CRs-Fixed: 2401151