OpenSSL: Drop support for OpenSSL 0.9.8
The OpenSSL project will not support version 0.9.8 anymore. As there
won't be even security fixes for this branch, it is not really safe to
continue using 0.9.8 and we might as well drop support for it to allow
cleaning up the conditional source code blocks.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>

Git-commit: de213e84e0e64472309f656b96b3fd64789f599e
Git-repo: git://w1.fi/srv/git/hostap.git
Change-Id: I5b0991ff95f2e881d658c0792a49221164881203
CRs-Fixed: 2373060