eap_proxy: Fix eap_proxy_init() prototype to use const eapol_cb
The eapol_cb structure was made const and that change resulted in a
compilation warning/error if CONFIG_EAP_PROXY=<name> is enabled in the
wpa_supplicant build configuration. Fix this by updating the function
prototype to match the change.

Note: This results in a change needed to external eap_proxy_*.c
implementations to match the change.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>
Git-commit: 79a54ab9f65448a9f3f8f149da5379614ad7bdca
Git-repo: git://w1.fi/srv/git/hostap.git
CRs-Fixed: 2322083
Change-Id: I18cd348ff9b27fcab7992f69bcb3c3d1c4cae7cb