P2P: Disallow GO CS immediately after GO Negotiation or invitation
A newly created GO might move to another channel before the client was
able to connect to it. This creates a situation where the client
searches the GO on the channel agreed upon during GO Negotiation or
invitation signaling, while the GO is on another channel. This in turn
might lead to delayed connection or connection failure and group
removal.

Fix this by disallowing a GO CS as long as there is some activity that
should delay the switch. If a GO move is not allowed, set a timeout to
re-attempt the move.

Signed-off-by: Ilan Peer <ilan.peer@intel.com>

Git-commit: dae4c82c9573268edd5dabc577b7bc7184b5ce22
Git-repo: git://w1.fi/srv/git/hostap.git
Change-Id: I2f4eaf0bbe5a58fc574d5498dc0f8beb7d619735
CRs-Fixed: 2270421