OWE: Improve discovery of OWE transition mode AP
An OWE AP device that supports transition mode does not transmit
the SSID of the OWE AP in its Beacon frames and in addition the
OWE AP does not reply to broadcast Probe Request frames. Thus,
the scan results matching relies only on Beacon frames from the
OWE open AP which can be missed in case the AP's frequency is
actively scanned.

To improve the discovery of transition mode APs, include their
SSID in the scan command to perform an active scan for the SSIDs
learned from the open mode BSSs.

Git-repo : git://w1.fi/srv/git/hostap.git
Git-commit: c04562e67edd7c2de6c0bf90c54ebfc24c7894ae
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
CRs-fixed: 2613441

Change-Id: Ice4d722f4dd86b7753a08edaae87fb0b4ce9e82f