HS 2.0: Fix Terms and Conditions URL handling
Do not try to determine the length of the T&C Server URL before checking
that the URL is available. This got broken in a change to move the
handling to the AS. hostapd could potentially have hit a NULL pointer
dereference if the authentication server sent an unconsistent set of T&C
information.

Fixes: d4e39c51f8bb ("HS 2.0: Move Terms and Conditions Server URL generation from AP to AS")
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>

Git-commit: ab3ab50757534114b1d005e19dd9d4b30cd5fab7
Git-repo: git://w1.fi/srv/git/hostap.git
Change-Id: Icb77395388f7a07525a88860a626f991b7061e10
CRs-Fixed: 2378773