EAP-pwd client: Verify received scalar and element
When processing an EAP-pwd Commit frame, the server's scalar and element
(elliptic curve point) were not validated. This allowed an adversary to
bypass authentication, and act as a rogue Access Point (AP) if the
crypto implementation did not verify the validity of the EC point.

Fix this vulnerability by assuring the received scalar lies within the
valid range, and by checking that the received element is not the point
at infinity and lies on the elliptic curve being used. (CVE-2019-9499)

The vulnerability is only exploitable if OpenSSL version 1.0.2 or lower
is used, or if LibreSSL or wolfssl is used. Newer versions of OpenSSL
(and also BoringSSL) implicitly validate the elliptic curve point in
EC_POINT_set_affine_coordinates_GFp(), enforcing security.

Git-repo : git://w1.fi/srv/git/hostap.git
Git-commit: 8ad8585f91823ddcc3728155e288e0f9f872e31a
Signed-off-by: Mathy Vanhoef <mathy.vanhoef@nyu.edu>
CRs-fixed: 2440806

Change-Id: I0e5f61cefce421c46b9f3d8a0a7dc065fb228189