SAE: Reuse previously generated PWE on a retry with the same AP
Do not start SAE authentication from scratch when the AP requests
anti-clogging token to be used. Instead, use the previously generated
PWE as-is if the retry is for the same AP and the same group. This saves
unnecessary processing on the station side in case the AP is under heavy
SAE authentiation load.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>

Git-repo : git://w1.fi/srv/git/hostap.git
Git-commit: fd830891202b56c41e26207ba6434ee352795935
CRs-fixed: 2414695
Change-Id: I0a518123a85809e77dd38124afe676a6ac147c7e