HE: Fix he_bss_color documentation
This field needs to be set to a value within 1-63 range, i.e., 0 is not
a valid value and does not indicate that BSS color is disabled. B7 of
the BSS Color octet is used to indicate that the BSS Color is
_temporarily_ disabled, but that is something that would happen
automatically based on detecting a collision in the used BSS colors and
not something that would be configured.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>

Git-repo : git://w1.fi/srv/git/hostap.git
Git-commit: b1664c8e4f00e0fe47d6fd26c1b67ee3bb1eb5b1
CRs-fixed: 2594511
Change-Id: Ibf1e90f6b3ad176709c2147d9cac318111a35b76