SAE: revert commit eb3234c07b9e0ae80e27c559dc5cc29feefaed87
For WFA case 5.2.1, if keep the previous change, the auth type set to host
drivet is open, but actually the test AP auth type is SAE. In some vendor
driver, it would compare the auth type with remote AP. If not match, will
return fail, then notify supplicant to delete pmkid cache. So case 5.2.1
will indicate pmkid is lost.

Change-Id: If16d55abd08d4aaf2d36930f2dc6838ad278b862