SAE: Use BSSID stored in ext_auth_bssid for set_pmk
pending_bssid is cleared in the connected state and thus is not
valid if SAE Authentication is done to a new BSSID when in
connected state. Hence refer the BSSID BSSID from ext_auth_bssid
while configuring the PMK. This is required for roaming to a new
BSSID on SME based drivers while the SAE processing happens with
wpa_supplicant.

CRs-fixed: 2510423
Change-Id: Ice3731e63bd00d42af6c120a9320d6378f19b7ee