nl80211: Don't set offchan-OK flag if doing on-channel frame in AP mode
I saw a case where the kernel's cfg80211 rejected hostapd's attempt to
send a neighbor report response because nl80211 flagged the frame as
offchannel-OK, but kernel rejects because channel was 100 (DFS) and so
kernel failed thinking it was constrained by DFS/CAC requirements that
do not allow the operating channel to be left (at least in FCC).

Don't set the packet as off-channel OK if we are transmitting on the
current operating channel of an AP to avoid such issues with
transmission of Action frames.

Signed-off-by: Ben Greear <greearb@candelatech.com>

Git-repo : git://w1.fi/srv/git/hostap.git
Git-commit: ff774311807ba9d72e0d92bc8a7eb118beb29346
CRs-fixed:2659327
Change-Id: I496d54520815bd425e1dbb45f130ed34f056c00f