eap_proxy: use eloop to spawn the thread for eap_proxy init
In the case of daemonize enabled, the parent process
that created the qmi client context will exit, and so
does the thread for initializing qmi client, which will
result in authentication failure, i.e. EAP-SIM.

To fix this issue, revert commit
982a34627ecbb845d9a9e76ee1554cfc8e72663c ("eap_proxy:
Avoid using eloop context for creating eap_proxy_post_init."),
and fix the issue mentioned in it by adding validation
check for eap_proxy when creating/starting the thread.

Change-Id: Ie3f3e73f1040e48e6825c676fda5d5d09b4e0575
CRs-Fixed: 2832589